Stream XRP and XLM transaction in unified network connection module
Introduces some common metadata fields in the top level tx object and
updates the tx filtering and graphical components to work with both
network transactions.

XRP transactions are no longer converted when streamed as when
these objects are wrapped the 'transaction' and 'meta' fields will
reside under the top level 'transaction' object as intended.

(When a tx is pulled in individually via /tx it is converted as
 it is in a different format).

Ziti will need to be updated to incorporate this change.